pyproject.toml
[project]
name = "benzema216-council"
version = "0.1.0"
description = "BENZEMA216 Council — Swarm intelligence investment council with 21 legendary investors"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
b216 = "benzema216_council.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/benzema216_council"]
